see entire commit message.
I am experimenting with a new structure for organizing my dotfiles.
The new structure allows for a folder to be created containing any
dotfiles minus the '.'(dot), e.g. .vimrc -> vimrc, which will
automatically be picked up and linked with bootstrap.sh.

usage to create a new category and link:

```
$ mkdir category
$ cp bootstrap.sh.template category/bootstrap.sh
$ mv ~/.dotfile category/dotfile
$ ./install.sh
```

If a .dotfile exists and is not already a symbolic link, it should
automatically be backed up in a directory called 'backups' next to the
install script.

Future plans include:
    - automatic generation of a category directory and creating a copy
      of the bootstrap template.
    - ability to 'uninstall' and restore backed up .dotfiles
    - automatic patching of the bootstrap.sh scripts